Rewrote axis as ES6 module.
