+++ es6_modules/rollup.entry.js
@@ -1,31 +1,32 @@
-// This file is only the entry point for rollup
-import {
-    ChartInternal
-} from './chartinternal.js';
-import {
-    Chart
-} from './chart.js';
-import {
-    Axis
-} from './axis.js';
+/**
+ * C3.js
+ * (c) 2016 Masayuki Tanaka and the C3.js contributors (See package.json)
+ * License: MIT
+ */
+
+import { ChartInternal } from './chartinternal.js';
+import { Chart } from './chart.js';
+import { Axis } from './axis/index';
 
 
-var version = "0.4.11",
-    generate = function(config) {
-        return new Chart(config);
+const version = '0.4.11';
+
+const generate = function generate(config) {
+    return new Chart(config);
+};
+
+const chart = {
+    fn: Chart.prototype,
+    internal: {
+        fn: ChartInternal.prototype,
+        axis: {
+            fn: Axis.prototype,
+        },
     },
-    chart = {
-        fn: Chart.prototype,
-        internal: {
-            fn: ChartInternal.prototype,
-            axis: {
-                fn: Axis.prototype
-            }
-        }
-    };
+};
 
 export {
     version,
     generate,
-    chart
-}
+    chart,
+};
